+++ src/domain/bootstrap/rootSaga.ts
@@ -1,5 +1,6 @@
 import { all, fork } from "redux-saga/effects";
 
+import appSaga from "../app/state/saga";
 import mapSaga from "../map/state/saga";
 import serviceSaga from "../service/saga";
 import unitSaga from "../unit/state/saga";
@@ -7,6 +8,7 @@ import searchSaga from "../unit/state/search/saga";
 
 export default function* rootSaga() {
   yield all([
+    fork(appSaga),
     fork(unitSaga),
     fork(searchSaga),
     fork(mapSaga),
