+++ src/domain/unit/state/saga.ts
@@ -1,13 +1,8 @@
 import { schema } from "normalizr";
 import { all, call, fork, put, takeLatest } from "redux-saga/effects";
 
-import {
-  callApi,
-  createRequest,
-  normalizeEntityResults,
-  stringifyQuery,
-} from "../../api/apiHelpers";
-import { FetchAction, Action } from "../../app/appConstants";
+import { callApi, normalizeEntityResults } from "../../api/apiHelpers";
+import { FetchAction } from "../../app/appConstants";
 import {
   NormalizedUnit,
   Unit,
@@ -42,31 +37,6 @@ function* clearSearch() {
   yield put(receiveSearchSuggestions([]));
 }
 
-function* sendFeedback({ payload: { feedback, email } }: Action) {
-  const params = {
-    description: feedback,
-    service_request_type: "OTHER",
-    can_be_published: false,
-    internal_feedback: true,
-    service_code: 2815,
-    email: undefined,
-  };
-
-  if (email) {
-    params.email = email;
-  }
-
-  const request = createRequest("https://api.hel.fi/servicemap/open311/", {
-    method: "POST",
-    body: stringifyQuery(params),
-    headers: {
-      "content-type": "application/x-www-form-urlencoded",
-    },
-  });
-
-  yield call(callApi, request);
-}
-
 function* watchFetchUnits() {
   yield takeLatest(UnitActions.FETCH, fetchUnits);
 }
@@ -75,14 +45,6 @@ function* watchClearSearch() {
   yield takeLatest(UnitActions.SEARCH_CLEAR, clearSearch);
 }
 
-function* watchSendFeedback() {
-  yield takeLatest(UnitActions.SEND_FEEDBACK, sendFeedback);
-}
-
 export default function* saga(): any {
-  yield all([
-    fork(watchFetchUnits),
-    fork(watchClearSearch),
-    fork(watchSendFeedback),
-  ]);
+  yield all([fork(watchFetchUnits), fork(watchClearSearch)]);
 }
